requirements.txt
fastapi==0.109.0
sqlmodel==0.0.16
uvicorn==0.25.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
alembic==1.13.1
pytest==7.4.4
loguru==0.7.2  
python-jose[cryptography]==3.3.0